move admin ui groups into individual components
